% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_to_lowd.R
\name{ComparisonPipeline}
\alias{ComparisonPipeline}
\title{Comparison Pipeline}
\usage{
ComparisonPipeline(orig, lowd, input_markers, k_titration)
}
\arguments{
\item{orig}{Tibble of cells by features}

\item{lowd}{Tibble of dimr output}

\item{input_markers}{Vector of strings of names of surface markers of interest}

\item{k_titration}{Vector of values corresponding to the number of nearest
neighbors k to be tried. We recommend a range from very small (eg. 5) up to
half the dataset.}

\item{lowd_markers}{Vector of strings corresponding to the names of the lowd
embedding columns in your cells tibble.}
}
\value{
A tibble of cells by k values, where each value is a given cell's
local fidelity relative to the k value of interst.
}
\description{
Loop to generate KNN from original space, and lowd space (or
any other input), and determine how well the tSNE and PCA space approximate
the orignal manifold.
}
\examples{
library(dplyr)
k.titration <- c(10, 100)
tsne_names <- names(samusik_tsne)
ComparisonPipeline(samusik_cells, samusik_surface_markers, samusik_tsne, k.titration)
}
