% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dqvis_cells}
\alias{dqvis_cells}
\title{Publically available dataset from Andreas Gruetzkau at Deutsches Rheuma
Forschung Zentrum (DRFZ)
Flow Repository ID FR-FCM-ZZWC}
\format{a tibble of 10000 cells by 49 features}
\usage{
dqvis_cells
}
\description{
Publically available dataset from Andreas Gruetzkau at Deutsches Rheuma
Forschung Zentrum (DRFZ)
Flow Repository ID FR-FCM-ZZWC
}
\keyword{datasets}
