% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dqvis_tsne}
\alias{dqvis_tsne}
\title{The output of t-SNE results from the dqvis_cells dataset with the
dqvis_surface_markers used as input.}
\format{a tibble of 10000 cells by 2 features}
\usage{
dqvis_tsne
}
\description{
The output of t-SNE results from the dqvis_cells dataset with the
dqvis_surface_markers used as input.
}
\keyword{datasets}
