% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nnvis_surface_markers}
\alias{nnvis_surface_markers}
\title{Selected surface markers from the dqvis_cells dataset to be used for dim
reduction analysis}
\format{a vector of 29 features}
\usage{
nnvis_surface_markers
}
\description{
Selected surface markers from the dqvis_cells dataset to be used for dim
reduction analysis
}
\keyword{datasets}
