% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_to_lowd.R
\name{RunPca}
\alias{RunPca}
\title{Run Principal Components Analysis (PCA)}
\usage{
RunPca(cells, input)
}
\arguments{
\item{cells}{Tibble of cells by features}

\item{input}{Vector of markers to be considered as input for the PCA method}
}
\value{
A tibble of cells by PC1 and PC2
}
\description{
Wrapper for 2-dimensional PCA
}
\examples{
data(wand.il7, package = "Sconify")
data(markers, package = "Sconify")
input <- markers[[1]]
RunPca(wand.il7, input)
}
