% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_to_lowd.R
\name{RunTsne}
\alias{RunTsne}
\title{Run t-SNE}
\usage{
RunTsne(cells, input, perp = 30)
}
\arguments{
\item{cells}{Tibble of cells by features}

\item{input}{Vector of markers to be considerd as input for the t-SNE method}

\item{perp}{The perplexity for t-SNE. Set to the CyTOF default of 30}
}
\value{
A tibble of cells by t-SNE1 and t-SNE2
}
\description{
Wrapper for 2-dimensional t-SNE
}
\examples{
data(wand.il7, package = "Sconify")
data(markers, package = "Sconify")
input <- markers[[1]]
RunTsne(wand.il7, input)
}
