% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_to_lowd.R
\name{RunUmap}
\alias{RunUmap}
\title{Runs UMAP on one's cells}
\usage{
RunUmap(cells, input = names(cells))
}
\arguments{
\item{cells}{TIbble of cells by features}

\item{input}{Vector of markers to be considered as input for the UMAP method}
}
\value{
A tibble of cells by UMAP1 and UMAP2
}
\description{
Wrapper for 2-dimensional UMAP
}
\examples{
RunUmap(dqvis_cells[1:1000,])
}
