% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_to_lowd.R
\name{RunVae}
\alias{RunVae}
\title{Run variational autoencoder}
\usage{
RunVae(cells, input = names(cells), epochs = 50L)
}
\arguments{
\item{cells}{Tibble of cells by features}

\item{input}{Vector of markers to be considered as input for the VAE}

\item{epochs}{the number of epochs for autoencoder training.}
}
\value{
A tibble of cells by enc1 and enc2
}
\description{
Wrapper for simple variational autoencoder through keras
}
\examples{
RunVae(dqvis_cells[1:1000,])
}
