% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samusik_cells}
\alias{samusik_cells}
\title{Publically available dataset from Nikolay Samusik of the Nolan Lab,
consisting of mouse bone marrow. Original data is 86,864 cells, but here
it is subsampled down to 10,000. The data were asinh transformed. The
Flow Repository ID FR-FCM-ZZPH. File is called Samusik_01.fcs.}
\format{a tibble of 10000 cells by 54 features}
\usage{
samusik_cells
}
\description{
Publically available dataset from Nikolay Samusik of the Nolan Lab,
consisting of mouse bone marrow. Original data is 86,864 cells, but here
it is subsampled down to 10,000. The data were asinh transformed. The
Flow Repository ID FR-FCM-ZZPH. File is called Samusik_01.fcs.
}
\keyword{datasets}
