% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_to_lowd.R
\name{RunTsne}
\alias{RunTsne}
\title{Run t-SNE}
\usage{
RunTsne(cells, input, perp = 30)
}
\arguments{
\item{cells}{Tibble of cells by features}

\item{input}{Vector of markers to be considerd as input for the t-SNE method}
}
\value{
A tibble of cells by t-SNE1 and t-SNE2
}
\description{
Wrapper for 2-dimensional t-SNE
}
