% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_to_lowd.R
\name{RunUMAP}
\alias{RunUMAP}
\title{Run UMAP}
\usage{
RunUMAP(cells, input = names(cells))
}
\arguments{
\item{cells}{Tibble of cells by features}

\item{input}{Vector of markers to be considered as input for UMAP method}
}
\value{
A tibble of cells by UMAP1 and UMAP2
}
\description{
Wrapper for 2-dimensional UMAP
}
\examples{
RunUMAP(samusik_cells[1:1000,], samusik_surface_markers)
}
